split gmp command into single files
split commands into single files:
- filters
- report config
- report format
