+++ src/gmp/commands/__tests__/report-config.test.ts
@@ -4,7 +4,7 @@
  */
 
 import {describe, test, expect} from '@gsa/testing';
-import {ReportConfigCommand} from 'gmp/commands/report-configs';
+import {ReportConfigCommand} from 'gmp/commands/report-config';
 import {
   createHttp,
   createEntityResponse,
