+++ src/gmp/commands/filters.ts
@@ -6,13 +6,10 @@
 import CollectionCounts from 'gmp/collection/collection-counts';
 import {type CollectionList, parseCollectionList} from 'gmp/collection/parser';
 import EntitiesCommand from 'gmp/commands/entities';
-import EntityCommand from 'gmp/commands/entity';
 import type Http from 'gmp/http/http';
 import {type XmlResponseData} from 'gmp/http/transform/fast-xml';
-import logger from 'gmp/log';
 import Filter, {type FilterModelElement} from 'gmp/models/filter';
 import type {Element} from 'gmp/models/model';
-import {resourceType, type EntityType} from 'gmp/utils/entity-type';
 import {isArray, isDefined} from 'gmp/utils/identity';
 
 interface FilterCountElement {
@@ -31,71 +28,10 @@ interface FiltersResponseElement extends Element {
   filter_count?: FilterCountElement;
 }
 
-interface GetFilterResponseData extends XmlResponseData {
-  get_filter?: {
-    get_filters_response?: {
-      filter?: FilterModelElement;
-    };
-  };
-}
-
 interface GetFiltersResponseData extends XmlResponseData {
   get_filters?: {
     get_filters_response?: FiltersResponseElement;
   };
-}
-
-const log = logger.getLogger('gmp.commands.filters');
-
-export class FilterCommand extends EntityCommand<Filter, FilterModelElement> {
-  constructor(http: Http) {
-    super(http, 'filter', Filter);
-  }
-
-  create(args: {
-    term: string;
-    name: string;
-    type: EntityType;
-    comment?: string;
-  }) {
-    const {term, name, type, comment = ''} = args;
-    const data = {
-      cmd: 'create_filter',
-      term,
-      name,
-      resource_type: resourceType(type),
-      comment,
-    };
-    log.debug('Creating new filter', args, data);
-    return this.action(data);
-  }
-
-  save(args: {
-    id: string;
-    term: string;
-    name: string;
-    type: EntityType;
-    comment?: string;
-  }) {
-    const {id, term, name, type, comment = ''} = args;
-    const data = {
-      cmd: 'save_filter',
-      comment,
-      id,
-      name,
-      resource_type: resourceType(type),
-      term,
-    };
-    log.debug('Saving filter', args, data);
-    return this.action(data);
-  }
-
-  getElementFromRoot(root: XmlResponseData): FilterModelElement {
-    return (
-      (root as GetFilterResponseData).get_filter?.get_filters_response
-        ?.filter ?? ({} as FilterModelElement)
-    );
-  }
 }
 
 const isPaginationElement = (
@@ -162,3 +98,5 @@ export class FiltersCommand extends EntitiesCommand<Filter> {
     };
   }
 }
+
+export default FiltersCommand;
